# docs/swagger.yaml
basePath: /
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.CreateNewFacultyRequest:
    properties:
      faculty_name:
        example: FITIP
        type: string
    required:
    - faculty_name
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.FacultyInfoResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      university_name:
        type: string
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User'
      user_roles:
        items:
          type: string
        type: array
    type: object
...
  github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.User:
    properties:
      first_name:
        example: John
        type: string
      id:
        example: 123456789
        type: integer
      language_code:
        example: ru
        type: string
      last_name:
        example: Doe
        type: string
      photo_url:
        example: https://example.com/photo.jpg
        type: string
      username:
        example: johndoe
        type: string
    required:
    - first_name
    - id
    type: object
...
  internal_http_handlers.RefreshRequest:
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
info:
  contact: {}
  description: API for Hackaton backend
  title: Hackaton API
  version: "1.0"
paths:
...
  /admin/faculties:
    get:
      consumes:
      - application/json
      description: Get all faculties for the university associated with the authenticated
        admin user
      produces:
      - application/json
      responses:
        "200":
          description: List of faculties
          schema:
            items:
              $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.FacultyInfoResponse'
            type: array
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden - user is not admin
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Get all faculties for admin's university
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: Create a new faculty for the university. Admin role required.
      parameters:
      - description: Faculty data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.CreateNewFacultyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: faculty created successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized user
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: Forbidden - user is not admin
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Create new faculty
      tags:
      - admin
...
  /auth/checkToken:
    get:
      consumes:
      - application/json
      description: Verify if the provided JWT access token is valid
      produces:
      - application/json
      responses:
        "200":
          description: 'status: token is valid'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - BearerAuth: []
      summary: Check JWT token validity
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authenticate user using MAX WebApp init data and return JWT tokens
      parameters:
      - description: Authentication date
        in: formData
        name: auth_date
        required: true
        type: string
      - description: Authentication hash
        in: formData
        name: hash
        required: true
        type: string
      - description: User data JSON
        in: formData
        name: user
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.LoginResponse'
        "400":
          description: Invalid request data
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Invalid init data
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: User login via MAX WebApp
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access and refresh tokens using a valid refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_http_handlers.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New JWT tokens
          schema:
            $ref: '#/definitions/github_com_max-main-team_backend_hackaton_MAX_internal_http_dto.LoginResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Refresh JWT tokens
      tags:
      - auth
...
swagger: "2.0"